Integrate duplicate parts of item_map.py and calculate.py into item.py and add function.
The common part of item_map.py and calculate has been moved to item.py.
When an item is input, the options and sub-items of the item are print. In the case of the lowest item, the farm location is also print.
