+++ items.json
@@ -81,7 +81,7 @@
     "108404": [["타구봉", "Mallet"], 3, [108301, 401303], "AP +85,AS +50%,HN +40%,", []],
     "108501": [["스파이의 우산", "Spy Umbrella"], 3, [108401, 401210], "AP +65,SA +25,SR +0.8,", []],
     "108502": [["여의봉", "Monkey King Bar"], 4, [401403, 108202], "AP +135,AR +1,", []],
-    "104503": [["마법봉", "Magic Stick"], 3, [108202, 205401], "AP +60,SA +27%,LS +30%,", []],
+    "108503": [["마법봉", "Magic Stick"], 3, [108202, 205401], "AP +60,SA +27%,LS +30%,", []],
     "109101": [["채찍", "Whip"], 0, 0, "AP +12,", [[14, 3], [15, 4]]],
     "109201": [["오랏줄", "Rope Cuffs"], 1, [109101, 401109], "AP +24,SA +6,", []],
     "109202": [["철편", "Bullwhip"], 1, [109101, 113101], "AP +30,", []],
@@ -270,8 +270,8 @@
     "201408": [["월계관", "Laurel Wreath"], 4, [205403, 201202], "EN +21,MS +580,SR +2,DF +20,HR +2,", [[-3, 1]]],
     "201409": [["제국왕관", "Imperial Crown"], 3, [201401, 401214], "SA +8%,MS +500,DF +18,MH +300,", [[-3, 1]]],
     "201410": [["황실 부르고넷", "Imperial Burgonet"], 3, [201402, 401214], "SA +18%,CR +10%,MS +200,DF +10,SP +0.1,", [[-3, 1]]],
-    "201411": [["변검", "Chinese Opera Mask"], 3, [201201, 401401], "MS +200,DF +20,SO +0.1,LS +20%,AE 킬 사용 후 4초 이내의 다음 기본 공격은 레벨*7.5의 피해를 추가로 입힙니다.,", [[-3, 1]]],
-    "201412": [["천사의 고리", "Angel Halo", "천고"], 4, [201202, 401403], "AP +15,DF +30,MS +300,AL +8%,AE 스킬 사용 후 4초 이내의 다음 기본 공격은 대상 최대 체력의 6%의 피해를 추가로 입힙니다.,", [[-3, 1]]],
+    "201411": [["변검", "Chinese Opera Mask"], 3, [201201, 401401], "MS +200,DF +20,SO +0.1,LS +20%,AE ,|스킬 사용 후 4초 이내의 다음 기본 공격은 레벨*7.5의 피해를 추가로 입힙니다.", [[-3, 1]]],
+    "201412": [["천사의 고리", "Angel Halo", "천고"], 4, [201202, 401403], "AP +15,DF +30,MS +300,AL +8%,AE ,|스킬 사용 후 4초 이내의 다음 기본 공격은 대상 최대 체력의 6%의 피해를 추가로 입힙니다.", [[-3, 1]]],
     "202101": [["바람막이", "Windbreaker"], 0, 0, "MH +50,ND +3,", [[4, 7], [6, 7], [15, 7], [-1, 8]]],
     "202103": [["승복", "Monk's Robe"], 0, 0, "CR +5%,DF +5,", [[9, 9], [10, 6], [-1, 4]]],
     "202105": [["전신 수영복", "Wetsuit"], 0, 0, "DF +1,SD +5%,", [[1, 7], [3, 6], [5, 7], [-1, 8]]],
@@ -294,7 +294,7 @@
     "202402": [["석양의 갑옷", "Sunset Armor"], 2, [202201, 401204], "DF +18,MH +360,", [[-3, 1]]],
     "202404": [["어사의", "Covert Agent Uniform"], 2, [202306, 401109], "CR +12%,DF +30,HR +2,VR +2,", [[-3, 1]]],
     "202405": [["광학미채 슈트", "Optical Camouflage Suit", "광미채"], 3, [202210, 401305], "AP +10,CC +24%,DF +18,SD +10%,", [[-3, 1]]],
-    "202406": [["락커의 자켓", "Rocker's Jacket"], 3, [202301, 502206], "AP +28,DF +18,ND +13,SO +0.1,HS +40%(원거리 25%),", [[-3, 1]]],
+    "202406": [["락커의 자켓", "Rocker's Jacket"], 3, [202301, 502206], "AP +28,DF +18,ND +13,SO +0.1,HS ,|+40%(원거리 25%)", [[-3, 1]]],
     "202407": [["미스릴 갑옷", "Mithril Armor"], 3, [401304, 202201], "AS +12%,DF +60,SP +0.1,SD +14,", [[-3, 1]]],
     "202408": [["성기사의 갑옷", "Crusader Armor"], 3, [202305, 205202], "EN +15,DF +35,HR +150%,", [[-3, 1]]],
     "202410": [["아마조네스 아머", "Amazoness Armor"], 3, [202305, 202209], "DF +28,MH +155,SD +20%,", [[-3, 1]]],
@@ -302,7 +302,7 @@
     "202412": [["지휘관의 갑옷", "Commander's Armor"], 3, [202302, 401214], "SA +18%,DF +26,", [[-3, 1]]],
     "202413": [["집사복", "Butler's Suit"], 3, [202303, 205209], "AP +30,ND +10,SP +0.12,LS +20%,", [[-3, 1]]],
     "202415": [["배틀 슈트", "Battle Suit"], 3, [202401, 202210], "DF +25,MH +170,ND +15,SD +8%,", [[-3, 1]]],
-    "202416": [["불꽃 드레스", "Blazing Dress"], 4, [205403, 202207], "AP +20,SR +75%,SR +2,DF +30,HR +2,FB 근처 3m이내의 적에게 착용자 최대 체력의 2.5%(원거리 캐릭터는 2%)의 피해를 매 초 입힙니다.,", [[-3, 1]]],
+    "202416": [["불꽃 드레스", "Blazing Dress"], 4, [205403, 202207], "AP +20,SR +75%,SR +2,DF +30,HR +2,FB ,|전투중에 근처 3m이내의 적에게 착용자 최대 체력의 2.5%(원거리 캐릭터는 2%)의 피해를 매 초 입힙니다.", [[-3, 1]]],
     "202417": [["EOD 슈트", "EOD Suit", "이오디슈트"], 3, [202401, 202206], "CR +10%,DF +25,MH +250,HR +1,ND +10,", [[-3, 1]]],
     "202501": [["카바나", "Kabana"], 4, [401209, 202412], "SA +20%,DF +40,", [[-3, 1]]],
     "202502": [["퀸 오브 하트", "Queen of Hearts"], 4, [401401, 202306], "CR +25%,DF +40,HR +2,MR +10,", [[-3, 1]]],
@@ -320,7 +320,7 @@
     "203303": [["바주반드", "Bazuband"], 2, [203203, 401213], "DF +25,ND +8,", [[-3, 1]]],
     "203304": [["진홍 팔찌", "Crimson Bracelet"], 2, [203104, 502201], "MS +250,MH +200,", [[-3, 1]]],
     "203401": [["강철 방패", "Steel Shield"], 2, [203201, 401201], "AP +23,DF +20,SP -0.05,", [[-3, 1]]],
-    "203402": [["소드 스토퍼", "Sword Stopper"], 3, [203303, 401101], "DF +36,ND +12,HN +40%(원거리 25%),", [[-3, 1]]],
+    "203402": [["소드 스토퍼", "Sword Stopper"], 3, [203303, 401101], "DF +36,ND +12,HN ,|+40%(원거리 25%)", [[-3, 1]]],
     "203403": [["드라우프니르", "Draupnir"], 3, [203302, 203203], "SA +17%,MS +240,DF +15,HR +1,", [[-3, 1]]],
     "203404": [["미스릴 방패", "Mithril Shield"], 3, [401304, 203201], "AP +30,AS +12%,DF +30,SP +0.1,", [[-3, 1]]],
     "203405": [["바이탈 센서", "Vital Sign Sensor"], 3, [203101, 401399], "AS +35%,DF +15,", [[-3, 1]]],
@@ -350,7 +350,7 @@
     "204407": [["미스릴 부츠", "Mithril Boots"], 3, [401304, 204202], "AS +12%,DF +18,SP +0.45,", [[-3, 1]]],
     "204408": [["부케팔로스", "Bucephalus"], 3, [204405, 204202], "AP +10,CC +15%,SR +1,DF +10,SP +0.33,", [[-3, 1]]],
     "204409": [["EOD 부츠", "EOD Boots"], 4, [204302, 204401], "SR +0.5,DF +20,HR +1,SP +0.3,TR +8%,", [[-3, 1]]],
-    "204411": [["클링온 부츠", "White Rhinos"], 3, [204404, 401101], "AS +15%,HR +0.7,SP +0.3,HN +40%(원거리 25%),", [[-3, 1]]],
+    "204411": [["클링온 부츠", "White Rhinos"], 3, [204404, 401101], "AS +15%,HR +0.7,SP +0.3,HN ,|+40%(원거리 25%)", [[-3, 1]]],
     "204410": [["글레이셜 슈즈", "Glacial Shoes"], 4, [204301, 205402], "SA +22,SR +60%,SR +1.5,HR +1.5,SP +0.25,SO +0.35,", [[-3, 1]]],
     "204501": [["헤르메스의 부츠", "Boots of Hermes"], 4, [401403, 204204], "DF +35,SP +0.35,SO +0.1,TH +20%,", [[-3, 1]]],
     "204502": [["분홍신", "Red Shoes"], 4, [204405, 401401], "AP +5,CC +15%,SR +1,SP +0.35,LS +12%,", [[-3, 1]]],
@@ -372,13 +372,13 @@
     "205207": [["진신사리", "Buddha Sarira"], 1, [205106, 202103], "SA +8%,CR +8%,", [[-3, 1]]],
     "205208": [["화살통", "Quiver"], 1, [401103, 108103], "AS +25%,", [[-3, 1]]],
     "205209": [["먼지털이개", "Feather Duster"], 1, [108102, 205101], "AP +9,SP +0.1,", [[-1, 16], [-3, 1]]],
-    "205210": [["군선", "Gilded Quill Fan"], 1, [205105, 401101], "SA +12,HS +40%,", [[-3, 1]]],
+    "205210": [["군선", "Gilded Quill Fan"], 1, [205105, 401101], "SA +12,HS ,|+40%(원거리 25%)", [[-3, 1]]],
     "205211": [["비파단도", "Shaman's Bronze"], 1, [102101, 401109], "AP +5,SA +8%,", [[-3, 1]]],
     "205212": [["사격교본", "Shooting Manual"], 1, [115101, 117101], "AP +8,AR +0.35,", [[-3, 1]]],
     "205301": [["생명의 가루", "Powder of Life"], 2, [401208, 401198], "SR +1.5,HR +1.5,", [[-1, 96]]],
     "205302": [["우치와", "Uchiwa"], 2, [205205, 205105], "SA +23,SR +80%,LS +14%,", []],
     "205303": [["탄창", "Magazine"], 2, [205107, 401213], "MH +330,ND +6,", []],
-    "205201": [["백우선", "White Crane Fan"], 2, [205210, 205101], "SA +18,SP +0.1,HS +40%(원거리 25%),", []],
+    "205201": [["백우선", "White Crane Fan"], 2, [205210, 205101], "SA +18,SP +0.1,HS ,|+40%(원거리 25%)", []],
     "205401": [["달빛 펜던트", "Moonlight Pendant"], 2, [205103, 401301], "AP +24,LS +16%,AR +1.5,", []],
     "205304": [["궁기병의 화살통", "Laced Quiver"], 2, [205208, 205209], "AP +13,AS +30%,SP +0.1,", [[-3, 1]]],
     "205305": [["월왕구천", "Revenge of Goujian"], 2, [205211, 401212], "AP +15,SA +15%,", [[-3, 1]]],
